+++ thesis-app/src/dashboard/admin/Dashboard-Admin.html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="Dashboard-Admin-Style.css">
+    <title>Dashboard</title>
+</head>
+<body>
+    <!-- Navigation Bar -->
+    <nav class="navbar">
+        <div class="nav-left">
+            <img src="c:/Image/NEU.png" alt="Logo" class="logo">
+            <ul class="nav-links">
+                <li><a href="#">Home</a></li>
+                <li><a href="#" class="active">Dashboard</a></li>
+                <li><a href="#">Submissions</a></li>
+                <li><a href="#">References</a></li>
+                <li><a href="#">Archives</a></li>
+            </ul>
+        </div>
+        <div class="nav-right">
+            <div class="notification-icon">
+                <img src="c:/Image/Bell.png" alt="Bell Icon" class="bell-icon">
+            </div>
+            <div class="profile-icon">
+                <img src="c:/Image/User.png" alt="User Icon">
+                <button class="down-button">
+                    <img src="c:/Image/DownButton.png" alt="Down Arrow">
+                </button>
+            </div>
+        </div>
+    </nav>
+
+    <!-- Dashboard -->
+    <div class="dashboard-container">
+        <h1>Dashboard</h1>
+        <div class="search-bar">
+            <img src="c:/Image/SearchIcon.png" alt="Search Icon" class="search-icon"> 
+            <input type="text" placeholder="Search">
+        </div>
+    </div>
+</body>
+</html>
