+++ thesis-app/src/login/css/Dashboard-Student-Style.css
@@ -1,0 +1,129 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: Arial, sans-serif;
+}
+
+body {
+    background-color: #E8E8E8;
+}
+
+/* Navigation Bar */
+.navbar {
+    background-color: #fff;
+    padding: 10px 20px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    position: relative; 
+    z-index: 10; 
+}
+
+.nav-left {
+    display: flex;
+    align-items: center;
+}
+
+.navbar .logo {
+    height: 60px; 
+}
+
+.nav-links {
+    list-style: none;
+    display: flex;
+    gap: 20px;
+    margin-left: 15px;
+}
+
+.nav-links a {
+    text-decoration: none;
+    color: #333;
+    font-weight: normal; 
+}
+
+.nav-links a.active {
+    font-weight: bold;
+}
+
+.nav-right {
+    display: flex;
+    align-items: center;
+    gap: 15px;
+}
+
+.nav-right a {
+    text-decoration: none;
+    color: #333;
+}
+
+.profile-icon {
+    display: flex;
+    align-items: center;
+    gap: 5px; 
+}
+
+.profile-icon img {
+    height: 30px;
+    border-radius: 50%;
+}
+
+.down-button {
+    background: none;
+    border: none;
+    padding: 0;
+    display: inline-block;
+    line-height: 0; 
+    cursor: pointer;
+}
+
+.down-button img {
+    width: 16px;  
+    height: 16px;
+}
+
+.down-button:hover img {
+    opacity: 0.8; 
+}
+
+.bell-icon {
+    width: 15; 
+    height: 15px; 
+    cursor: pointer; 
+}
+
+/* Dashboard Text and Search Bar */
+.dashboard-container {
+    text-align: center;
+    padding: 50px;
+    margin-top: 50px; 
+}
+
+.dashboard-container h1 {
+    font-size: 36px;
+    margin-bottom: 20px;
+    margin-top: 20px; 
+}
+
+.search-bar {
+    position: relative;
+    display: inline-flex;
+    align-items: center;
+    margin-top: 20px; 
+}
+
+.search-bar input[type="text"] {
+    width: 300px;
+    padding: 10px 10px 10px 35px; 
+    border: none;
+    border-radius: 20px;
+    outline: none;
+}
+
+.search-bar .search-icon {
+    position: absolute;
+    left: 10px;
+    width: 20px; 
+    height: 20px; 
+    pointer-events: none; 
+}
