+++ thesis-app/src/login/css/Login.css
@@ -1,0 +1,149 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: 'Poppins', sans-serif;
+}
+ 
+body {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-height: 100vh;
+    background: #dfdfdf;
+}
+
+.login-box {
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    width: 580px;
+    padding: 30px;
+    background: rgba(255, 255, 255, 0.9); 
+    border-radius: 15px;
+    box-shadow: 0px 5px 15px rgba(0, 0, 0, 0.2);
+}
+
+.login-header {
+    text-align: center;
+    margin: 20px 0 30px; 
+}
+
+.login-header header {
+    color: #333;
+    font-size: 30px;
+    font-weight: 600;
+}
+
+.input-box .input-field {
+    width: 100%;
+    height: 60px;
+    font-size: 17px;
+    padding: 0 25px;
+    margin-bottom: 15px; 
+    border-radius: 30px;
+    border: none;
+    box-shadow: 0px 5px 10px rgba(0, 0, 0, 0.05);
+    outline: none;
+    transition: .3s;
+}
+
+::placeholder {
+    font-weight: 500;
+    color: #222;
+}
+
+.input-field:focus {
+    width: 100%; 
+}
+
+.forgot {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 30px;
+}
+
+a {
+    text-decoration: none;
+    color: #555;
+}
+
+a:hover {
+    text-decoration: underline;
+}
+
+.input-submit {
+    position: relative;
+}
+
+.submit-btn {
+    width: 45%;
+    height: 60px;
+    background: #222;
+    border: none;
+    border-radius: 30px;
+    cursor: pointer;
+    transition: .3s;
+    margin: 0 auto;
+    display: block; 
+}
+
+.input-submit label {
+    position: absolute;
+    top: 50%; 
+    left: 50%;
+    color: #fff;
+    transform: translate(-50%, -50%);
+    cursor: pointer;
+}
+
+.submit-btn:hover {
+    background: #000;
+    transform: scale(1.05);
+}
+
+.or {
+    text-align: center;
+    color: #555;
+    margin-top: 10px;
+}
+
+.icons {
+    text-align: center; 
+}
+
+.input-icon {
+    width: 25px; 
+    height: 25px; 
+}
+
+.google-signin {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 45%; 
+    height: 60px; 
+    background: #fff; 
+    border: 2px solid #ccc; 
+    border-radius: 30px;
+    cursor: pointer; 
+    margin: 10px auto; 
+    transition: background 0.3s, transform 0.3s; 
+}
+
+.google-signin:hover {
+    background: #f5f5f5; 
+    transform: scale(1.05); 
+}
+
+.google-signin img {
+    margin-right: 10px; 
+}
+
+.terms {
+    text-align: center;
+    font-size: 14px;
+    margin-top: 25px; 
+    margin-bottom: 5x;
+}
